websocket testing...

+++ ESP32/Websocket_Client/src/EzoSensor.cpp
@@ -1,6 +1,6 @@
-#include <EzoSensor.h>
+#include "EzoSensor.hpp"
 
-EzoSensor::EzoSensor(uint8_t address, char * name) {
+EzoSensor::EzoSensor(int address, char * name) {
 	this->address = address;
 	this->name = name;
 }
